feat: Add value_stop parameter to Convergence_checker_optuna
This commit adds a new parameter, `value_stop`, to the `Convergence_checker_optuna` class in the `helpers.py` file. The `value_stop` parameter allows for stopping the optimization when the best value is less than or equal to the specified value. This change improves the control over the optimization process.

Also makes yaml_save maintain order